extend profiler API
Summary: Extend current profiler API. No op to existing profiler.cc since it is already broken and unused. Note that, the current profiler.cc will be replaced by a plugin and refactored in upstream. But the timeline is unknown. Thus, don't see a need to fix the default profiler.cc.

Reviewed By: wesbland

Differential Revision: D45597201

fbshipit-source-id: 912e992d90eec5a91dfb75417fe7e6cc3e66e6c9